Adapt PsiProvider for new signal syntax
